move to testmonitor client
