--- nisystemlink/clients/step/__init__.py
@@ -1,3 +1,0 @@
-from ._step_client import StepClient
-
-# flake8: noqa
