+++ nisystemlink/clients/testmonitor/models/_update_steps_partial_success.py
@@ -2,8 +2,8 @@ from typing import List, Optional
 
 from nisystemlink.clients.core import ApiError
 from nisystemlink.clients.core._uplink._json_model import JsonModel
-from nisystemlink.clients.step.models._step import Step
-from nisystemlink.clients.step.models._update_steps_request import UpdateStepRequestObject
+from nisystemlink.clients.testmonitor.models._step import Step
+from nisystemlink.clients.testmonitor.models._update_steps_request import UpdateStepRequestObject
 
 
 class UpdateStepsPartialSuccess(JsonModel):
